Refactor gradients.
Previously I strived to match the CSS spec when it came to gradients.
With use I realized that trying to mach the specc excactly is a waste of
time. This is supposed to be a opinionated library, so I decided to find
some opinions. Gradients now do away with many of the positional
keywords, at the same time there are no optional arguments. You need to
understand how gradients work and apply each parameter. It is way better
this way.